sam3_service/requirements-service.txt
# SAM3 服务层依赖
# 官方 sam3 仓库的依赖请参考 sam3/requirements.txt

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pillow>=10.0.0
numpy>=1.24.0
pydantic>=2.0.0
